feat: refactor user store to integrate auth store and improve user session handling

+++ src/stores/user.ts
@@ -1,27 +1,20 @@
-import { defineStore } from 'pinia'
-import { STORAGE_KEY } from './auth'
+import { defineStore, storeToRefs } from 'pinia'
 import type { components } from '@/types/api'
-import { ref, watch } from 'vue'
 import { http } from '@/services/http'
+import { useAuthStore } from '@/stores/auth'
 
-type UserDto = components['schemas']['UserDto']
+export type UserDto = components['schemas']['UserDto']
 type UsersPageDto = components['schemas']['UsersPageDto']
 type PartialUserDto = components['schemas']['PartialUserDto']
 
 export const useUserStore = defineStore('user', () => {
-  const userStored = localStorage.getItem(STORAGE_KEY)
-  const loggedUser = ref<UserDto | null>(userStored ? JSON.parse(userStored) : null)
+  const authStore = useAuthStore()
+  const { loggedUser } = storeToRefs(authStore)
 
-  watch(
-    loggedUser,
-    (val) => {
-      if (val) localStorage.setItem(STORAGE_KEY, JSON.stringify(val))
-      else localStorage.removeItem(STORAGE_KEY)
-    },
-    { immediate: true },
-  )
-
-  const usersByPagination = async (queries: { page: number; limit: number }) => {
+  const usersByPagination = async (queries: {
+    page: number
+    limit: number
+  }): Promise<UsersPageDto> => {
     const { data } = await http.get<UsersPageDto>('/users', {
       params: {
         ...queries,
@@ -31,23 +24,44 @@ export const useUserStore = defineStore('user', () => {
   }
 
   const partialLoggedUpdateUser = async (user: PartialUserDto): Promise<UserDto> => {
+    if (loggedUser.value === null) {
+      throw new Error('User not logged in')
+    }
+
     const { data } = await http.patch<UserDto>(`/users`, user)
     loggedUser.value = data
     return data
   }
 
   const deleteLoggedUser = async () => {
+    if (loggedUser.value === null) {
+      throw new Error('User not logged in')
+    }
+
     await http.delete(`/users`)
-    loggedUser.value = null
+    authStore.logout()
   }
 
   const getLoggedUser = async () => {
-    const { data } = await http.get<UserDto>(`/users/${loggedUser.value?.uuid}`)
+    if (loggedUser.value === null) {
+      throw new Error('User not logged in')
+    }
+
+    const { data } = await http.get<UserDto>(`/users/${loggedUser.value.uuid}`)
     loggedUser.value = data
     return data
   }
 
+  const userByUuid = async (uuid: string): Promise<UserDto> => {
+    const { data } = await http.get<UserDto>(`/users/${uuid}`)
+    return data
+  }
+
   const uploadProfilePicture = async (file: File) => {
+    if (loggedUser.value === null) {
+      throw new Error('User not logged in')
+    }
+
     const formData = new FormData()
     formData.append('file', file)
     const { data } = await http.post<UserDto>(`/users/profile-picture`, formData, {
@@ -60,6 +74,10 @@ export const useUserStore = defineStore('user', () => {
   }
 
   const uploadBannerPicture = async (file: File) => {
+    if (loggedUser.value === null) {
+      throw new Error('User not logged in')
+    }
+
     const formData = new FormData()
     formData.append('file', file)
     const { data } = await http.post<UserDto>(`/users/banner-picture`, formData, {
@@ -79,5 +97,6 @@ export const useUserStore = defineStore('user', () => {
     getLoggedUser,
     uploadProfilePicture,
     uploadBannerPicture,
+    userByUuid,
   }
 })
